Add Support for CELO, RSK and ETC
This commit adds support for non-ETH chains with concrete
implementations for CELO, RSK, and ETC. The main change for these chains
is that the operation hash prefix must be specific to the chain, to
avoid cross-chain replay attacks. For example, on ETH chain it is the
string "ETHER", on RSK chain it is the string "RSK" etc. In addition,
this commit extends the test suite to cover the wallet
implementations, and adds a test to ensure that signatures with the
wrong operatin hash prefix fail.

Ticket: BG-22389